LayoutTests/media/track/webvtt-ruby-expected.txt


EVENT(canplay)
EVENT(addtrack)
EXPECTED (video.textTracks.length == '1') OK
RUN(video.textTracks[0].mode = 'showing')
RUN(video.currentTime = 1)
EVENT(seeked)
EXPECTED (window.internals.shadowRoot(video).querySelector('rt') != 'null') OK
EXPECTED (rubyText.getBoundingClientRect().top < rubyBase.getBoundingClientRect().top == 'true') OK
EXPECTED (rubyBase.getBoundingClientRect().top == rubyText.getBoundingClientRect().bottom == 'true') OK
END OF TEST
